many web changes
   * refactoring to reduce clutter in web/mod.rs
   * some refactoring of request guards for convenience
   * new pages:
      * season history
      * per-season info
      * homepage
