+++ Cargo.toml
@@ -32,7 +32,7 @@ aliri_braid = "0.2.4"
 serde = {version = "1", features=["derive"]}
 serde_json = "1"
 chrono = "0.4"
-chrono-tz = "0.6.3"
+chrono-tz = "0.8.1"
 bb8 = "0.8.0"
 async-trait = "0.1"
 itertools = "0.10"
