+++ http/templates/base.html.tera
@@ -9,25 +9,27 @@
 	    {% endblock %}
 	</head>
 	<body class="bg-sky-400">
-	    <div class="bg-sky-600 h-7 mb-2">
-	        <div class="flex">
-	            <div class="flex md:w-1/5">
-	            </div>
-	            <div class="flex flex-auto">
-	                <a href="/brackets">Brackets</a>
-	            </div>
-
-	            <div class="flex flex-auto">
-	                <a href="/standings">Standings</a>
-	            </div>
-	            <div class="flex md:w-1/5">
-	            </div>
-	        </div>
+        <div class="bg-sky-600 h-7">
+            <div class="flex">
+                <div class="flex md:w-1/5">
+                </div>
+                <div class="flex flex-auto">
+                    <a href="/">home</a>
+                </div>
+                {% if base_context.current_season %}
+                <div class="flex flex-auto">
+                    <a href="/season/{{base_context.current_season.id}}">Current season</a>
+                </div>
+                {% endif %}
+                <div class="flex flex-auto">
+                    <a href="/seasons">history</a>
+                </div>
+                <div class="flex md:w-1/5">
+                </div>
+            </div>
             {# header #}
-	    </div>
-	    <div class="p-2 text-2xl md:text-xl lg:text-base">
+        </div>
 		{% block body %}
 		{% endblock %}
-		</div>
 	</body>
 </html>
