+++ http/templates/macros.tera
@@ -1,6 +1,6 @@
-{% macro link(href, text) %}
-    <a href="{{ href }}" class="underline text-emerald-600 visited:text-amethyst-600">{{ text }}</a>
-{% endmacro link %}
+{# for some reason, the -% here doesn't actually owrk, so the {% endmacro %} has to go on the same line :\ #}
+{%- macro link(href, text) -%}
+    <a href="{{ href }}" class="underline text-emerald-600 visited:text-amethyst-600">{{ text }}</a>{% endmacro link %}
 
 {% macro header(name) %}
 <th class="border-x border-slate-500">
