+++ http/templates/season_brackets.html.tera
@@ -1,12 +1,11 @@
-{% extends "base" %}
+{% extends "season_base" %}
 
 {% import "macros" as macros %}
 
-{% block pagename %}NMG League Season 2 Brackets{% endblock %}
-{% block body %}
-{% if season %}
-    <span title="Season #{{ season.id}}">{{ season.format }}</span>
-{% endif %}
+{# #TODO + 1 LOOOOOOOOOOOL god dammit fix this someday #}
+{% block pagename %}NMG League {{season.format}} Brackets{% endblock %}
+{% block season_body %}
+
 {% for bracket in brackets %}
     <div class="p1 border border-1">
         <span title="Bracket id#{{ bracket.bracket.id }}">Bracket: {{ bracket.bracket.name }}</span>
