+++ src/discord/bot.rs
@@ -26,7 +26,7 @@ use twilight_model::id::Id;
 use twilight_standby::Standby;
 use twilight_util::builder::InteractionResponseDataBuilder;
 
-use crate::constants::{APPLICATION_ID_VAR, TOKEN_VAR};
+use nmg_league_bot::constants::{APPLICATION_ID_VAR, TOKEN_VAR};
 use nmg_league_bot::db::get_diesel_pool;
 use crate::discord::discord_state::DiscordState;
 use crate::discord::interactions_utils::{
