+++ src/discord/mod.rs
@@ -8,7 +8,6 @@ use std::sync::Arc;
 use bb8::RunError;
 use chrono::{DateTime, Duration, TimeZone};
 use diesel::ConnectionError;
-use twilight_http::error::ErrorType;
 use twilight_http::request::channel::reaction::RequestReactionType;
 use twilight_http::Client;
 use twilight_mention::timestamp::{Timestamp as MentionTimestamp, TimestampStyle};
@@ -32,7 +31,6 @@ use nmg_league_bot::models::player::{MentionOptional, Player};
 use nmg_league_bot::models::race::Race;
 use nmg_league_bot::models::race_run::RaceRun;
 use nmg_league_bot::utils::{race_to_nice_embeds, ResultErrToString};
-use nmg_league_bot::{ChannelConfig, NMGLeagueBotError};
 use thiserror::Error;
 use nmg_league_bot::worker_funcs::{clear_commportunities_message, clear_tentative_commentary_assignment_message};
 pub(crate) use webhooks::Webhooks;
